// app/types/user.ts
// export type User = {
//   id: number;
//   name: string;
//   username: string;
//   email: string;
//   address: Address;
//   phone: string;
//   website: string;
//   company: Company;
// };

// export type Address = {
//   street: string;
//   suite: string;
//   city: string;
//   zipcode: string;
//   geo: Geo;
// };

// export type Company = {
//   name: string;
//   catchPhrase: string;
//   bs: string;
// };

// export type Geo = {
//   lat: string;
//   lng: string;
// };

export type Coordinates= {
  lat: number;
  lng: number;
}

export type Address= {
  address: string;
  city: string;
  coordinates: Coordinates;
  postalCode: string;
  state: string;
}

export type Hair= {
  color: string;
  type: string;
}

export type Bank= {
  cardExpire: string;
  cardNumber: string;
  cardType: string;
  currency: string;
  iban: string;
}

export type CompanyAddress= {
  address: string;
  city: string;
  coordinates: Coordinates;
  postalCode: string;
  state: string;
}

export type Company= {
  address: CompanyAddress;
  department: string;
  name: string;
  title: string;
}

export type User= {
  id: number;
  firstName: string;
  lastName: string;
  maidenName: string;
  age: number;
  gender: string;
  email: string;
  phone: string;
  username: string;
  password: string;
  birthDate: string;
  image: string;
  bloodGroup: string;
  height: number;
  weight: number;
  eyeColor: string;
  hair: Hair;
  domain: string;
  ip: string;
  address: Address;
  macAddress: string;
  university: string;
  bank: Bank;
  company: Company;
  ein: string;
  ssn: string;
  userAgent: string;
}